chore(s3): move streamToString to own file
